Implemented a few basic methods
Implemented the following methods. init to intialize the 2D array based on total classroom and large classroom constraints. fillLargeClasses to fill the newly vreated array with large classes in the earliest possible time slots. primaryGoal calculates the total timeslots used, which is our job to minimize as the primary goal of our model. secondaryGoal calculates the number of wasted classes, which is our secondary goal to minimize. I put in a method stub for beginIteration, as I have an idea for how this should work but didn't have time to start on before class. My idea is that the method starts with the initial array and we only put small classes in small classrooms, and then we calculate the primary and secondary goals of the completed array and store them. We then decrease the number of small classes in small classrooms by 1 and put that 1 class in a big classroom, calculate the goals, and if its better than the first pass through, store it. We continue this for all possible configurations of small classrooms and store the best configuration, like finding the maximum of a number set.